backend/partie-1/requirements.txt
# Backend Python – RAG Chatbot
# Installer avec : pip install -r requirements.txt

# Framework web
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20

# OpenAI (LLM + Embeddings)
openai>=1.0.0

# Extraction de texte PDF
pypdf>=4.0.0

# Extraction DOCX / PPTX
python-docx>=1.1.0
python-pptx>=1.0.0

# OCR images (Tesseract doit être installé séparément)
pytesseract>=0.3.10
Pillow>=10.0.0

# Comptage de tokens (chunking précis)
tiktoken>=0.7.0

# Variables d'environnement
python-dotenv==1.0.1

# Validation des données
pydantic>=2.0.0

# Base de données vectorielle (Supabase pgvector)
supabase==2.11.0

# Intégration Outlook (Microsoft Graph API)
msal>=1.28.0
httpx>=0.27.0
